requirements.txt
requests
logging
discord.py